Config tab page groundwork
Created tab scroll frame and scroll child
Began to add tab option creation.
Moved default settings to a separate file to better compartmentalise and
improve clarity of main file.
Reconfigured settings table and references to aid with the selection of
the right type of option selector.
Added copyright boilerplate to top of config and default settings